backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
numpy==2.1.3
python-dotenv==1.0.1
